feat: findAll Items, Categories
