+++ projects/client/src/pages/admin/order/Orders.jsx
@@ -1,12 +1,7 @@
-import Order from "../../../components/admin/order/Order";
 import Main from "../../../components/admin/main/Main";
 
 const Orders = () => {
-  return (
-    <Main title="Orders">
-      <Order />
-    </Main>
-  );
+  return <Main title="Orders"></Main>;
 };
 
 export default Orders;
