+++ projects/client/src/pages/admin/rack/Racks.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import Main from "../../../components/admin/main/Main";
 
 const Racks = () => {
